Upgrade to use BigInteger

+++ net.projecteuler.solutions/src/main/java/utils/Sequence.java
@@ -1,5 +1,7 @@
 package utils;
 
+import java.math.BigInteger;
+
 public abstract class Sequence {
-    public abstract long next();
+    public abstract BigInteger next();
 }
